+++ src/components/UploadDocumentCard/index.tsx
@@ -1,18 +1,16 @@
 // External Imports
 // React
-import React, { useState, useMemo, useCallback } from 'react';
+import React, { useMemo } from 'react';
 
 // Relative Imports
 // Components
 import { Icon } from '../Icon';
 import { Flex } from '../Flex';
 import { Typography } from '../Typography';
-import { Button } from '../Button';
-import { FlexCard, StyledLinearProgress } from './styles';
+import { FlexCard, StyledLinearProgress, LeftContent, Locked, Upload, Uploading, UploadCompleted } from './styles';
 
 export type UploadDocumentCardState = 'upload' | 'uploading' | 'uploaded' | 'error' | 'locked';
-
-type SharedProps = { label: string; fileName: string; fileSize?: string };
+export type SharedProps = { label: string; fileName: string; fileSize?: string };
 
 type CardWithMediaProps = SharedProps & { fileUrl: string };
 type RemovableMediaProps = CardWithMediaProps & {
@@ -46,36 +44,20 @@ const UploadDocumentCard = <State extends UploadDocumentCardState>(props: Upload
   const isLocked = useMemo(() => state === 'locked', [state]);
   const hasFile = useMemo(() => ['locked', 'uploaded', 'uploading'].includes(state), [state]);
 
-  const onUpload = useCallback((onSelect: (file: File) => void) => {
-    let input = document.createElement('input');
-    input.type = 'file';
-    input.accept = '.pdf';
-    input.onchange = function () {
-      if (input.files) {
-        onSelect(input.files?.[0]);
-      }
-    };
-    input.click();
-  }, []);
-
-  const onRemoveWrapper = useCallback((onRemove: () => void, event: React.MouseEvent) => {
-    event.stopPropagation();
-    onRemove();
-  }, []);
-
   // Render
-
-  // if (state === 'upload' || state === 'error' || state === 'uploaded' || state === 'uploading') {
-  //   // const onClick =
-  //   //   state === 'upload' ? onUpload.bind(null, (props as UploadDocumentCardProps<typeof state>).onSelect) : null;
-  // }
   const { label, fileName, fileSize } = props as UploadDocumentCardProps<typeof state>;
+  const { errorMessage } = props as UploadDocumentCardProps<'error'>;
+
   return (
     <Flex direction="column">
       <Typography class="medium" size="base" color="dark90">
         {label}
       </Typography>
-      {/* error text ??? */}
+      {errorMessage && (
+        <Typography class="medium" size="sm" color="error75">
+          {errorMessage}
+        </Typography>
+      )}
       <FlexCard direction="row" align="center" state={state}>
         <Icon
           icon={hasFile ? 'File' : 'CloudUpload'}
@@ -88,10 +70,10 @@ const UploadDocumentCard = <State extends UploadDocumentCardState>(props: Upload
             {/* left content */}
             <LeftContent state={state} fileName={fileName} fileSize={fileSize} />
             {/* right content */}
-            {(state === 'upload' || state === 'error') && <Upload />}
-            {state === 'uploading' && <Uploading />}
-            {state === 'uploaded' && <UploadComplete />}
-            {state === 'locked' && <Locked fileSize={fileSize} />}
+            {(state === 'upload' || state === 'error') && <Upload {...(props as UploadDocumentCardProps<'upload'>)} />}
+            {state === 'uploading' && <Uploading {...(props as UploadDocumentCardProps<'uploading'>)} />}
+            {state === 'uploaded' && <UploadCompleted {...(props as UploadDocumentCardProps<'uploaded'>)} />}
+            {props.state === 'locked' && <Locked {...(props as UploadDocumentCardProps<'locked'>)} />}
           </Flex>
           {/* linear progress */}
           {state === 'uploading' && (
@@ -107,93 +89,3 @@ const UploadDocumentCard = <State extends UploadDocumentCardState>(props: Upload
 };
 
 export { UploadDocumentCard };
-
-type LeftContentProps = {
-  state: UploadDocumentCardState;
-} & Pick<SharedProps, 'fileName' | 'fileSize'>;
-
-const LeftContent = ({ state, fileName, fileSize }: LeftContentProps) => {
-  const isError = state === 'error';
-  return (
-    <Flex direction="column" grow="1">
-      <Typography class="medium" size="sm" padding={0.4} color={isError ? 'error75' : 'dark75'}>
-        {fileName ?? 'Select a file or drag and drop here'}
-      </Typography>
-      {state !== 'locked' && (
-        <Typography class="roman" size="xs" padding={0.4} color={isError ? 'error60' : 'dark60'}>
-          {fileSize ?? 'maximum file size is 200mb'}
-        </Typography>
-      )}
-    </Flex>
-  );
-};
-
-type LockedProps = { fileSize: SharedProps['fileSize'] };
-
-const Locked: React.FC<LockedProps> = ({ fileSize }) => {
-  return (
-    <Flex direction="column" justify="center" alignSelf="stretch">
-      <Typography color="dark60" class="roman" size="xs" padding={0}>
-        {fileSize}
-      </Typography>
-    </Flex>
-  );
-};
-
-// type LeftContentUploadProps = {};
-
-const Upload: React.FC = () => {
-  return (
-    <Flex direction="row" align="center">
-      <Button
-        primaryVariant="secondary"
-        secondaryVariant="mint"
-        leadingIcon="Upload"
-        title="Upload file"
-        size="small"
-      />
-    </Flex>
-  );
-};
-
-type UploadingProps = {
-  uploadProgress?: number;
-};
-
-const Uploading: React.FC<UploadingProps> = ({ uploadProgress = 0 }) => {
-  return (
-    <Flex direction="row" align="center">
-      <Typography class="roman" size="sm" color="dark75" padding={0.4}>
-        {uploadProgress}%
-      </Typography>
-      <Icon icon="X" color="dark75" size={18} />
-    </Flex>
-  );
-};
-
-const UploadComplete: React.FC = () => {
-  const [confirmDelete, setConfirmDelete] = useState<boolean>(false);
-
-  const toggleConfirmDelete = useCallback(() => setConfirmDelete(curr => !curr), []);
-
-  if (!confirmDelete) {
-    return (
-      <Flex direction="column" justify="center" alignSelf="stretch">
-        <span onClick={toggleConfirmDelete}>
-          <Icon icon="Trash" size={18} color="dark90" />
-        </span>
-      </Flex>
-    );
-  }
-
-  return (
-    <Flex direction="row" align="center" justify="center">
-      <div onClick={toggleConfirmDelete}>
-        <Typography class="roman" size="xs" color="dark60" padding={0}>
-          Cancel
-        </Typography>
-      </div>
-      <Button primaryVariant="primary" secondaryVariant="destructive" title="Delete" size="small" />
-    </Flex>
-  );
-};
